Base images should rarely expose ports
Given that it is unlikely that the alpine-nginx image would be deployed
as-is, it should not expose ports. Instead that should be left to the
image that uses alpine-nginx as its base.

This is especially true because, unlike `CMD` or `ENTRYPOINT` a derived
image _cannot_ override the `EXPOSE` command of a base image. This means
that should a derived image want nginx to _not_ expose ports 80 or 443,
they cannot use this as a base image.

Also, nginx logs should always be redirected to stdout and stderr, and
therefore the commands to make this happen should remain the
responsibility of alpine-nginx base image.